+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `@distributedlab/tools`: `BN` `isEqualTo` compare
+
+### Changed
+- `@distributedlab/tools`: `BN` comparing methods
+
+### Fixed
+- `@distributedlab/tools`: `BN` return formatted string for value getter instead toString method
 
 ## [0.1.5] - 2023-03-09
 ### Fixed